Converted 3rd person character into a first person character (see description)
A fully animated character that is using Blend Trees is used as the "body" of the player. A camera, that acts as the "eyes" of the player, uses a simple script to make sure it stays in the same position (onn the eyes of the character level). Player can rotate the camera to look around (rotation is limited; a human cannot rotate its head 180 degree, same applies here). A video with this is provided on Discord. Or if you played Outlast, it's the same perspective, only it uses better transitions for the animations.